Fix types and add complex test

+++ src/DataTypes/DataTypesNumber.cpp
@@ -46,8 +46,8 @@ void registerDataTypeNumbers(DataTypeFactory & factory)
     factory.registerDataType("Int16", createNumericDataType<Int16>);
     factory.registerDataType("Int32", createNumericDataType<Int32>);
     factory.registerDataType("Int64", createNumericDataType<Int64>);
-    factory.registerDataType("Float32", createNumericDataType<Int32>);
-    factory.registerDataType("Float64", createNumericDataType<Int64>);
+    factory.registerDataType("Float32", createNumericDataType<Float32>);
+    factory.registerDataType("Float64", createNumericDataType<Float64>);
 
     /// These synonyms are added for compatibility.
